Rewrote program based on moviepy library
I was earlier trying to use ffmpeg to generate the various videos. I made
the decision to switch to a pure python library as ffmpeg suffered from
way too many disadvantage in the context I wanted to use.
1. It was platform dependant,I would have had to write two set of ffmpeg
commands, one for windows and one for unix machines.
2. It has a steep learning curve that is much longer than the time I am
ready to spend on this project.
4. A user would have required ffmepg to be installed.
5. ffmpeg reduces code readability drastically.
6. I had no need for a extremely sophisticated video editing tool as my
needs were limited.